+++ parser_library/src/context/literal_pool.cpp
@@ -210,9 +210,6 @@ void literal_pool::generate_pool(diagnosable_ctx& diags, index_t<using_collectio
             diags.add_diagnostic(diagnostic_op::error_E033(it->second.r));
         else
         {
-            // some types require operands that consist only of one symbol
-            (void)lit->check_single_symbol_ok(diagnostic_collector(&diags, lit_val.stack));
-
             ord_ctx.symbol_dependencies().add_postponed_statement(
                 std::make_unique<literal_postponed_statement>(lit, lit_val),
                 dependency_evaluation_context {