+++ parser_library/src/context/ordinary_assembly/dependable.h
@@ -20,6 +20,9 @@
 
 #include "dependency_collector.h"
 
+namespace hlasm_plugin::parser_library {
+struct asm_option;
+}
 namespace hlasm_plugin::parser_library::expressions {
 struct data_definition;
 } // namespace hlasm_plugin::parser_library::expressions
@@ -46,6 +49,8 @@ public:
         id_index label, const section* owner, int32_t offset, bool long_offset) const = 0;
     virtual std::variant<const symbol*, symbol_candidate> get_symbol_candidate(id_index name) const = 0;
     virtual std::string get_opcode_attr(id_index symbol) const = 0;
+    virtual const asm_option& get_options() const noexcept = 0;
+    virtual const section* get_section(id_index name) const noexcept = 0;
 
 protected:
     ~dependency_solver() = default;