+++ parser_library/src/context/ordinary_assembly/dependency_solver_redirect.cpp
@@ -48,4 +48,11 @@ std::string dependency_solver_redirect::get_opcode_attr(id_index symbol) const
     return m_base->get_opcode_attr(symbol);
 }
 
+const asm_option& dependency_solver_redirect::get_options() const noexcept { return m_base->get_options(); }
+
+const section* dependency_solver_redirect::get_section(id_index name) const noexcept
+{
+    return m_base->get_section(name);
+}
+
 } // namespace hlasm_plugin::parser_library::context