+++ parser_library/src/context/ordinary_assembly/ordinary_assembly_dependency_solver.h
@@ -56,6 +56,8 @@ public:
         id_index label, const section* owner, int32_t offset, bool long_offset) const override;
     std::variant<const symbol*, symbol_candidate> get_symbol_candidate(id_index name) const override;
     std::string get_opcode_attr(id_index name) const override;
+    const asm_option& get_options() const noexcept override;
+    const section* get_section(id_index name) const noexcept override;
 
     dependency_evaluation_context derive_current_dependency_evaluation_context() const&;
     dependency_evaluation_context derive_current_dependency_evaluation_context() &&;